svg detection

+++ graphormer/data/algos.pyx
@@ -74,7 +74,7 @@ def gen_edge_input(max_dist, path, edge_feat):
     assert path_copy.flags['C_CONTIGUOUS']
     assert edge_feat_copy.flags['C_CONTIGUOUS']
 
-    cdef numpy.ndarray[long, ndim=4, mode='c'] edge_fea_all = -1 * numpy.ones([n, n, max_dist_copy, edge_feat.shape[-1]], dtype=numpy.int64)
+    cdef numpy.ndarray[float, ndim=4, mode='c'] edge_fea_all = -1 * numpy.ones([n, n, max_dist_copy, edge_feat.shape[-1]], dtype=numpy.float32)
     cdef unsigned int i, j, k, num_path, cur
 
     for i in range(n):
